+++ app/src/main/java/com/example/bmta/model/Item.kt
@@ -4,9 +4,9 @@ data class Item (
     override var name: String = "Item",
     override var position: Position,
     var pickedUp : Boolean = false,
-    val health: Float = 0.0F,
-    val attack: Float= 0.0F,
-    val defense : Float= 0.0F,
-    val healing : Float= 0.0F,
+    val health: Double = 0.0,
+    val attack: Double= 0.0,
+    val defense : Double= 0.0,
+    val healing : Double= 0.0,
     override var command: String = name.lowercase()
 ) : GameObject()
